Add files via upload
In this Project, I have visualized and analyzed data on the job outcomes of students
who graduated from college between 2010 and 2012. We find that student
employment tended to be lower in majors related to the humanities, arts, and
education and that these majors tended to have a higher percentage of women
enrolled. We also find that women's participation is highly volatile across majors,
with shares of as little as 2% in majors related to Engineering and Science.